--- Python/libraries/recognizers-date-time/recognizers_date_time/date_time/base_timezone.py
@@ -1,30 +1,0 @@
-from abc import ABC, abstractmethod
-from recognizers_text.matcher.string_matcher import StringMatcher
-from typing import List, Pattern
-
-
-class TimeZoneExtractorConfiguration(ABC):
-    @property
-    @abstractmethod
-    def direct_utc_regex(self) -> Pattern:
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def location_time_suffix_regex(self) -> Pattern:
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def location_matcher(self) -> StringMatcher:
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def timezone_matcher(self) -> StringMatcher:
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def ambiguous_timezone_list(self) -> List[str]:
-        raise NotImplementedError
